+++ test/log.txt
@@ -1,0 +1,2 @@
+Tue Oct 10 2017 13:43:24 GMT-0700 (PDT)  **  client 1 hello world
+Tue Oct 10 2017 13:43:24 GMT-0700 (PDT)  **  client 2 hello world
